prometheus-grafana/grafana-alloy/otel-demo/app/requirements.txt
Flask

# Core OpenTelemetry
opentelemetry-api==1.24.0
opentelemetry-sdk==1.24.0

# Instrumentation
opentelemetry-instrumentation==0.45b0
opentelemetry-instrumentation-flask==0.45b0
opentelemetry-instrumentation-logging==0.45b0

# Exporters
opentelemetry-exporter-otlp==1.24.0
opentelemetry-exporter-otlp-proto-http==1.24.0

